fix: multisig `executed` flag + rename filter util
